Oops, fixed some signed/unsigned comparisons

+++ src/TinyGsmClientBG96.h
@@ -556,7 +556,7 @@ protected:
     }
     int len = stream.readStringUntil('\n').toInt();
 
-    for (size_t i=0; i<len; i++) {
+    for (int i=0; i<len; i++) {
       TINY_GSM_MODEM_STREAM_TO_MUX_FIFO_WITH_DOUBLE_TIMEOUT
     }
     waitResponse();
